requirements.txt
pytest
PyYAML
pydantic
typer
pandas
pyarrow
rank-bm25
numpy
faiss-cpu
matplotlib
scikit-learn
